overlapping event windows don't always wipe each other out now
instead, they now carry on if they have the best (lowest) RNG score in the area. Sometimes they still wipe each other out if they are worse than other windows that are also wiped out. They don't recusively check this - i reckon that might be a nightmare and/or very slow to process. But who knows - maybe there's a better way of doing it.